+++ elkm1_lib/panel.py
@@ -28,7 +28,6 @@ class Panel(Element):
         self._elk.add_handler("XK", self._xk_handler)
         self._elk.add_handler("RR", self._xk_handler)  # RR/XK use same handler
         self._elk.add_handler("RP", self._rp_handler)
-        self._elk.add_handler("IE", self._elk.call_sync_handlers)
         self._elk.add_handler("SS", self._ss_handler)
         self._elk.add_handler("UA", self._ua_handler)
